Ensure that re-openned SHM have the same configuration as it did the first time.

+++ collections/src/main/java/net/openhft/collections/SharedMapErrorListener.java
@@ -17,7 +17,7 @@
 package net.openhft.collections;
 
 public interface SharedMapErrorListener {
-    void onLockTimeout(int threadId) throws IllegalStateException;
+    void onLockTimeout(long threadId) throws IllegalStateException;
 
     void errorOnUnlock(IllegalMonitorStateException e);
 }
