+++ collections/src/main/java/net/openhft/collections/SharedMapErrorListeners.java
@@ -22,7 +22,7 @@ import java.util.logging.Logger;
 public enum SharedMapErrorListeners implements SharedMapErrorListener {
     LOGGING {
         @Override
-        public void onLockTimeout(int threadId) throws IllegalStateException {
+        public void onLockTimeout(long threadId) throws IllegalStateException {
             Logger.getLogger(getClass().getName()).severe("Grabbing lock held by threadId: " + threadId);
         }
 
@@ -32,7 +32,7 @@ public enum SharedMapErrorListeners implements SharedMapErrorListener {
         }
     }, ERROR {
         @Override
-        public void onLockTimeout(int threadId) throws IllegalStateException {
+        public void onLockTimeout(long threadId) throws IllegalStateException {
             throw new IllegalStateException("Unable to acquire lock held by threadId: " + threadId);
         }
